\name{calcres-methods}

\docType{methods}

\alias{calcres-methods}
\alias{calcres,ANY-method}
\alias{calcres,system-method}

\title{\code{calcres}}

\description{
\code{\link{calcres}} calculates a dynamic solution of the model defined by the argument.
}

\section{Methods}{
\describe{

\item{\code{signature(system = "ANY")}}{
}

\item{\code{signature(system = "system")}}{
}
}}
