\name{calcsens-methods}

\docType{methods}

\alias{calcsens-methods}
\alias{calcsens,ANY-method}
\alias{calcsens,system-method}

\title{\code{calcsens}}

\description{
\code{\link{calcsens}} Performs a sensitivity analysis of the model passed as the argument.
}

\section{Methods}{
\describe{

\item{\code{signature(system = "ANY")}}{
}

\item{\code{signature(system = "system")}}{
}
}}
