\name{plotres}

\alias{plotres}

\title{
Plot Simulation Results
}

\description{
Produces a simple standardized plot of the results generated by \code{\link{calcres}} 
or, more generally, 
plots of all or selected columns of a matrix as a function of the row names interpreted as numbers.
}

\usage{
plotres(res, colnames = list(), file = "", ...)
}

\arguments{
  \item{res}{
Numerical matrix all or selected columns of which are to be plotted as a function of the row names.
}
  \item{colnames}{
Selection of column names to be plotted. An empty list indicates that all columns should be plotted.
}
  \item{file}{
Optional file name to which the plot should be redirected. The file will be written in pdf format.
}
  \item{\dots}{
Further arguments are transferred to the underlying plot routine.
}
}

\author{
Peter Reichert <peter.reichert@eawag.ch>
}

\seealso{
\code{\link{process-class}},
\code{\link{reactor-class}}
\code{\link{link-class}}
\code{\link{system-class}}
\code{\link{calcres}}.
}

\examples{
# see
help(ecosim)
# for examples.
}

